音频播放进度条与真是进度同步(还未实现拖动修改进度)

+++ app/src/main/java/com/niyangup/himalaya/interfaces/IPlayerViewCallback.java
@@ -30,7 +30,7 @@ public interface IPlayerViewCallback {
      */
     void onPlayModeChange(XmPlayListControl.PlayMode mode);
 
-    void onProgressChange(long currentProgress, long totalProgress);
+    void onProgressChange(int currentProgress, int totalProgress);
 
     void onAdLoading();
